+++ src/terms/CollectionsGrid.vue
@@ -3,6 +3,10 @@ import { ref, watchEffect } from "vue";
 import type { QlitCollection } from "@/services/qlit.types";
 import { getCollections } from "@/services/terms.service";
 
+defineProps<{
+  selected: string[];
+}>();
+
 const emit = defineEmits<{
   select: [QlitCollection];
 }>();
@@ -20,7 +24,10 @@ watchEffect(async () => (collections.value = await getCollections()));
       class="w-1/2 sm:w-1/3 p-1"
     >
       <div
-        class="p-2 px-4 bg-amber-300 shadow rounded-lg h-full flex justify-center items-center cursor-pointer"
+        class="p-2 px-4 bg-amber-300 hover:brightness-105 transition-all shadow rounded-lg h-full flex justify-center items-center cursor-pointer"
+        :class="{
+          'bg-amber-400 shadow-lg': selected?.includes(collection.name),
+        }"
         @click="emit('select', collection)"
       >
         {{ collection.label }}