+++ src/views/Thesaurus.vue
@@ -9,6 +9,7 @@ import useTitle from "@/views/title.composable";
 import TermTree from "@/terms/TermTree.vue";
 import LoadingSpinner from "@/components/LoadingSpinner.vue";
 import type { QlitTerm } from "@/services/qlit.types";
+import { getCollection } from "@/services/terms.service";
 
 const PAGE_SIZE = 15;
 
@@ -21,6 +22,7 @@ const termTextQuery = computed(() => state.termTextQuery);
 const limit = ref(PAGE_SIZE);
 const termsLimited = computed(() => terms.value.slice(0, limit.value));
 const isLoading = ref(false);
+const heading = ref<string>();
 
 onMounted(async () => {
   isLoading.value = true;
@@ -28,16 +30,28 @@ onMounted(async () => {
   isLoading.value = false;
 });
 
+// A debounced function for loading terms matching text input
 const findTerms = debounce(async () => {
   isLoading.value = true;
   terms.value = await searchTerms(termTextQuery.value);
   isLoading.value = false;
 }, 400);
 
+// Reactively determine what terms to show
 watchEffect(async () => {
   if (termTextQuery.value) {
+    // A text query has been entered
+    heading.value = `Ämnen på "${termTextQuery.value}"`;
     findTerms();
+  } else if (state.termCollection) {
+    // A collection is selected
+    heading.value = state.termCollection.label;
+    isLoading.value = true;
+    terms.value = await getCollection(state.termCollection.name);
+    isLoading.value = false;
   } else {
+    // Default case: show root terms
+    heading.value = "Ämnesträd";
     terms.value = rootTerms.value;
   }
 });
@@ -57,6 +71,7 @@ function onBottomVisibility(visible: boolean) {
   <div class="container py-6">
     <LoadingSpinner v-if="isLoading" />
     <div v-else-if="terms.length" :key="termTextQuery">
+      <h1 v-if="heading" class="text-4xl mb-6">{{ heading }}</h1>
       <TermTree
         v-for="term in termsLimited"
         :key="term.name"